+++ scripts/build-all.sh
@@ -10,5 +10,7 @@ ${DIR}/build.sh --version 1.11.1 --arch cpu
 ${DIR}/build.sh --version 1.11.1 --arch gpu
 ${DIR}/build.sh --version 1.12.0 --arch cpu
 ${DIR}/build.sh --version 1.12.0 --arch gpu
+${DIR}/build.sh --version 1.13.0 --arch cpu
+${DIR}/build.sh --version 1.13.0 --arch gpu
 ${DIR}/build.sh --version 1.11 --arch eia
 ${DIR}/build.sh --version 1.12 --arch eia