+++ test/integration/local/test_container.py
@@ -35,10 +35,10 @@ def volume():
         subprocess.check_call('docker volume rm model_volume'.split())
 
 
-@pytest.fixture(scope='module', autouse=True)
-def container(enable_batch, docker_base_name, tag):
+@pytest.fixture(scope='module', autouse=True, params=[True, False])
+def container(request, docker_base_name, tag):
     try:
-        if enable_batch:
+        if request.param:
             batching_config = ' -e SAGEMAKER_TFS_ENABLE_BATCHING=true'
         else:
             batching_config = ''
@@ -240,3 +240,17 @@ def test_predict_no_custom_attributes_header():
     y = json.loads(response.content.decode('utf-8'))
 
     assert y == {'predictions': [3.5, 4.0, 5.5]}
+
+
+def test_predict_with_jsonlines():
+    x = {
+        'instances': [1.0, 2.0, 5.0]
+    }
+
+    headers = {
+        'Content-Type': 'application/json',
+        'Accept':  'application/jsonlines'
+    }
+    response = requests.post(BASE_URL, data=json.dumps(x), headers=headers)
+    assert response.headers['Content-Type'] == 'application/jsonlines'
+    assert response.content.decode('utf-8') == '{    "predictions": [3.5, 4.0, 5.5    ]}'