+++ test/integration/sagemaker/util.py
@@ -16,8 +16,11 @@ import logging
 import os
 
 import botocore
+import random
+import time
 
 logger = logging.getLogger(__name__)
+BATCH_CSV = os.path.join('test', 'data', 'batch.csv')
 
 
 def image_uri(registry, region, repo, tag):
@@ -29,7 +32,7 @@ def _execution_role(boto_session):
 
 
 @contextlib.contextmanager
-def sagemaker_model(region, boto_session, sagemaker_client, image_uri, model_name, model_data):
+def sagemaker_model(boto_session, sagemaker_client, image_uri, model_name, model_data):
     model = sagemaker_client.create_model(
         ModelName=model_name,
         ExecutionRoleArn=_execution_role(boto_session),
@@ -58,14 +61,19 @@ def _production_variants(model_name, instance_type, accelerator_type):
 
     return production_variants
 
+
+def _test_bucket(region, boto_session):
+    account = boto_session.client('sts').get_caller_identity()['Account']
+    return f'sagemaker-{region}-{account}'
+
+
 def find_or_put_model_data(region, boto_session, local_path):
     model_file = os.path.basename(local_path)
 
-    account = boto_session.client('sts').get_caller_identity()['Account']
-    bucket = f'sagemaker-{region}-{account}'
+    bucket = _test_bucket(region, boto_session)
     key = f'test-tfs/{model_file}'
 
-    s3 = boto_session.client('s3')
+    s3 = boto_session.client('s3', region)
 
     try:
         s3.head_bucket(Bucket=bucket)
@@ -96,6 +104,11 @@ def find_or_put_model_data(region, boto_session, local_path):
 def sagemaker_endpoint(sagemaker_client, model_name, instance_type, accelerator_type=None):
     logger.info('creating endpoint %s', model_name)
 
+    # Add jitter so we can run tests in parallel without running into service side limits.
+    delay = round(random.random()*5, 3)
+    logger.info('waiting for {} seconds'.format(delay))
+    time.sleep(delay)
+
     production_variants = _production_variants(model_name, instance_type, accelerator_type)
     sagemaker_client.create_endpoint_config(EndpointConfigName=model_name,
                                             ProductionVariants=production_variants)
@@ -117,6 +130,85 @@ def sagemaker_endpoint(sagemaker_client, model_name, instance_type, accelerator_
         sagemaker_client.delete_endpoint_config(EndpointConfigName=model_name)
 
 
+def _create_transform_job_request(model_name, batch_output, batch_input, instance_type):
+    return {
+        'TransformJobName': model_name,
+        'ModelName': model_name,
+        'BatchStrategy': 'MultiRecord',
+        'TransformOutput': {
+            'S3OutputPath': batch_output
+        },
+        'TransformInput': {
+            'DataSource': {
+                'S3DataSource': {
+                    'S3DataType': 'S3Prefix',
+                    'S3Uri': batch_input
+                }
+            },
+            'ContentType': 'text/csv',
+            'SplitType': 'Line',
+            'CompressionType': 'None'
+        },
+        'TransformResources': {
+            'InstanceType': instance_type,
+            'InstanceCount': 1
+        }
+    }
+
+
+def _read_batch_output(region, boto_session, bucket, model_name):
+    s3 = boto_session.client('s3', region)
+    output_file = f'/tmp/{model_name}.out'
+    s3.download_file(bucket, f'output/{model_name}/batch.csv.out', output_file)
+    return json.loads(open(output_file, 'r').read())['predictions']
+
+
+def _wait_for_transform_job(region, boto_session, sagemaker_client, model_name, poll, timeout):
+    status = sagemaker_client.describe_transform_job(TransformJobName=model_name)['TransformJobStatus']
+    job_runtime = 0
+    while status == 'InProgress':
+
+        logger.info(f'Waiting for batch transform job {model_name} to finish')
+        time.sleep(poll)
+        job_runtime += poll
+        if job_runtime > timeout:
+            raise ValueError(f'Batch transform job {model_name} exceeded maximum runtime {timeout} seconds')
+
+        status = sagemaker_client.describe_transform_job(TransformJobName=model_name)['TransformJobStatus']
+        if status == 'Completed':
+            return _read_batch_output(region=region,
+                                      boto_session=boto_session,
+                                      bucket=_test_bucket(region, boto_session),
+                                      model_name=model_name)
+        if status == 'Failed':
+            raise ValueError(f'Failed to execute batch transform job {model_name}')
+        if status in ['Stopped', 'Stopping']:
+            raise ValueError(f'Batch transform job {model_name} was stopped')
+
+
+def run_batch_transform_job(region, boto_session, model_data, image_uri,
+                            sagemaker_client, model_name,
+                            instance_type):
+
+    with sagemaker_model(boto_session, sagemaker_client, image_uri, model_name, model_data):
+        batch_input = find_or_put_model_data(region, boto_session, BATCH_CSV)
+        bucket = _test_bucket(region, boto_session)
+        batch_output = f's3://{bucket}/output/{model_name}'
+
+        request = _create_transform_job_request(
+            model_name=model_name, batch_input=batch_input,
+            batch_output=batch_output, instance_type=instance_type)
+        sagemaker_client.create_transform_job(**request)
+
+    return _wait_for_transform_job(
+               region=region,
+               boto_session=boto_session,
+               sagemaker_client=sagemaker_client,
+               model_name=model_name,
+               poll=5, # seconds
+               timeout=300) # seconds
+
+
 def invoke_endpoint(sagemaker_runtime_client, endpoint_name, input_data):
     response = sagemaker_runtime_client.invoke_endpoint(EndpointName=endpoint_name,
                                                         ContentType='application/json',
@@ -126,9 +218,9 @@ def invoke_endpoint(sagemaker_runtime_client, endpoint_name, input_data):
     return result
 
 
-def create_and_invoke_endpoint(region, boto_session, sagemaker_client, sagemaker_runtime_client,
+def create_and_invoke_endpoint(boto_session, sagemaker_client, sagemaker_runtime_client,
                                model_name, model_data, image_uri, instance_type, accelerator_type,
                                input_data):
-    with sagemaker_model(region, boto_session, sagemaker_client, image_uri, model_name, model_data):
+    with sagemaker_model(boto_session, sagemaker_client, image_uri, model_name, model_data):
         with sagemaker_endpoint(sagemaker_client, model_name, instance_type, accelerator_type):
             return invoke_endpoint(sagemaker_runtime_client, model_name, input_data)